* refactor - add opt fp8 quantize kernel

 * fix - opt quantize and fix bug

 * fix - lite fix gang server

 * fix - fix alltoall fp8

 * fix: complict

 * fix: fix fp8 moe test

 * fix: [internal] comment out biencoder_tbstars_embedding
Link: https://code.alibaba-inc.com/foundation_models/FasterTransformer/codereview/21109630
* refactor - add opt fp8 quantize kernel

* fix - opt quantize and fix bug

* fix - lite fix gang server

* fix - fix alltoall fp8

* fix: complict

* fix: fix fp8 moe test

* fix: macro def and test

+++ src/fastertransformer/devices/cuda_impl/tests/CudaQuantizeTest.cc
@@ -14,7 +14,7 @@ class CudaQuantizeTest : public DeviceTestBase {
 public:
     void RunCudaQuantizeFp8() {
         int m = 128, n = 256;
-        auto input = torch::randn({m, n}).to(torch::kFloat32).to(torch::kCUDA);
+        auto input = torch::randn({m, n}).to(torch::kBFloat16).to(torch::kCUDA);
         auto output = device_->quantize({*torchTensor2Buffer(input),
                                          DataType::TYPE_FP8_E4M3,
                                          1,
